add in ability to update saved geo and herbicide app rates on search activity page
