+++ app/src/components/form/FormControlsComponent.tsx
@@ -1,5 +1,6 @@
 import { Button, Grid } from '@material-ui/core';
 import React from 'react';
+import { createNamedExports } from 'typescript';
 
 export interface IFormControlsComponentProps {
   classes?: any;
@@ -32,37 +33,30 @@ const FormControlsComponent: React.FC<IFormControlsComponentProps> = (props) => 
               Check Form For Errors
             </Button>
           </Grid>
-          <Grid item>
-            <Button
-              disabled={isDisabled}
-              variant="contained"
-              color="primary"
-              onClick={() => {
-                if (!props.onCopy) {
-                  return;
-                }
-
-                props.onCopy();
-              }}>
-              Copy Form Data
-            </Button>
-          </Grid>
+          {props.onCopy && (
+            <Grid item>
+              <Button
+                disabled={isDisabled}
+                variant="contained"
+                color="primary"
+                onClick={() => props.onCopy()}
+              >
+                Copy Form Data
+              </Button>
+            </Grid>
+          )}
           {
             sessionStorage.getItem('copiedFormData') &&
             sessionStorage.getItem('activitySubtype') === props.activitySubtype &&
+            props.onPaste &&
             (
               <Grid item>
                 <Button
                   disabled={isDisabled}
                   variant="contained"
                   color="primary"
-                  onClick={() => {
-                    if (!props.onPaste) {
-                      return;
-                    }
-
-                    props.onPaste();
-                  }}>
+                  onClick={() => props.onPaste()}
+                >
                   Paste Form Data
                 </Button>
               </Grid>
